requirements.txt
certifi==2021.10.8
click==8.0.3
cmake==3.22.2
dlib==19.23.0
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.22.1
opencv-python==4.5.5.62
Pillow==9.0.0
